Moving testing to Gulp. Not in working state.
