+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "test": "istanbul cover node_modules/mocha/bin/_mocha test/index.js && mocha --reporter supersamples test/integration/index.js",
+    "test2": "istanbul cover node_modules/mocha/bin/_mocha test/index.js && mocha --reporter supersamples test/integration/index.js",
     "coveralls": "npm run test -- --report lcovonly && cat ./coverage/lcov.info | coveralls"
   },
   "dependencies": {
@@ -35,12 +35,15 @@
     "eslint-config-google": "^0.7.1",
     "eslint-plugin-react": "^6.9.0",
     "gulp": "^3.9.1",
+    "gulp-connect": "^5.0.0",
     "gulp-eslint": "^3.0.1",
+    "gulp-mocha": "^3.0.1",
     "gulp-sass-lint": "^1.3.2",
     "istanbul": "^0.4.5",
     "mocha": "^3.2.0",
     "node-sass": "^4.4.0",
     "sass-loader": "^4.1.1",
+    "should": "^11.2.0",
     "style-loader": "^0.13.1",
     "supersamples": "^1.2.0",
     "supertest": "^3.0.0",
