+++ routes/api/action/action.route.spec.js
@@ -1,7 +1,5 @@
 const config = require('config');
 const request = require('supertest');
-const app = require('../../app');
-require('chai').should();
 
 describe('Action Route', function() {
     it('Create Action One', function(done) {
