--- test/index.js
@@ -1,2 +1,0 @@
-require('./unit/index');
-require('./integration/index');
