--- test/unit/index.js
@@ -1,3 +1,0 @@
-// Unit Tests
-require('./user.controller.test.js');
-require('./action.controller.test.js');
